refactor: merge distributed.go and child_executor.go
